Add files via upload
I made the motor gears higher (2mm in both directions) so if the gears wobble or it is not mounted precisely its not a big problem. Should make building and working with the hangprinter easier. I also made a Version where you can use the round motor shaft. If the gears have a tight fit on the shaft this should be enough to hold the gears.